requirements.txt
# ========================================
# Tech Challenge Fase 3 - Dependências
# Assistente Virtual Médico com LLM Fine-tuning
# ========================================

# Core ML/LLM
transformers>=4.36.0
torch>=2.1.0
datasets>=2.16.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0

# LangChain e LangGraph
langchain>=0.1.0
langchain-community>=0.0.10
langgraph>=0.0.20
langchain-openai>=0.0.5

# Processamento de dados
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0

# NLP e Tokenização
sentencepiece>=0.1.99
tiktoken>=0.5.0

# Anonimização de dados
faker>=22.0.0
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0

# Logging e Monitoramento
loguru>=0.7.0
wandb>=0.16.0
tensorboard>=2.15.0

# Utilitários
python-dotenv>=1.0.0
tqdm>=4.66.0
pyyaml>=6.0.0
click>=8.1.0

# API e Validação
pydantic>=2.5.0
fastapi>=0.108.0
uvicorn>=0.25.0

# Testes
pytest>=7.4.0
pytest-cov>=4.1.0

# Jupyter (opcional, para notebooks)
jupyter>=1.0.0
ipykernel>=6.27.0
